refactor: create a HeroCard component
